video: s5p-dp: retry dp_enable when Link Training is failed
This patch retries dp_enable when Link Training is failed.
In this case, LCD power pin should be off and on in order to
reset LCD panel. Therefore, callbacks such as lcd_on() and
lcd_off() are added to s5p_dp_platdata.

The retry count is set as 3 times, because it works properly
within 3 times.
Bug : 7185117
Change-Id: I193af78ffba644b24ef15a7c2fdfa2f8041ecee6
Signed-off-by: minwoo7945.kim <minwoo7945.kim@samsung.com>
Signed-off-by: Jingoo Han <jg1.han@samsung.com>