drm/exynos5: add irq resource to hdmi-audio device for hpd handling
Hdmi-audio device in alsa-soc needs hpd notifications for enabling and
disabling the audio. If hdmi-audio dapm switch is on and hdmi cable is
plugged-in, audio rendered on the speakers should also be routed to hdmi
audio sink. This patch adds irq resource information to the hdmi-audio
device so that it can requeste the hpd irq in shared mode.

BUG=chrome-os-partner:12908
TEST:verified hpd plug-unplug scenarions with audio playback and
dapm switch control toggling.

Change-Id: I3457e9c00d949205eca624395834ac92b4eb4bee
Signed-off-by: Rahul Sharma <rahul.sharma@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/33423
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Ready: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>