media: exynos: mfc: remove context buffer initialization
Context buffer initialization is removed because MFC F/W does it
internally. Also, do not clean the cache entries, but invalidate
them instead.

This fixes a problem during mediaserver restart. The current
context is pointing to a DRM context which we
absolutely cannot write and clean() to. So, we need to make sure
that we invalidate anything at that address so that it doesn't
get cleaned out later. Also, since f/w does the memset 0, we
need to make sure that stale data in the cache doesn't get
flushed out.

Bug: 7361993
Change-Id: Ic27d32f972124f535955aa3207a431d4d486c2a7
Signed-off-by: Janghyuck Kim <janghyuck.kim@samsung.com>
[ dima: changed for both ctx and instance bufs and inval only ]
Signed-off-by: Dima Zavin <dima@android.com>