ARM: EXYNOS: Fix the vsync interrupt timeout.
This patch fixes the vsync interrupt timeout.
The EXYNOS5250 is that set as '0' to CMU_SYSCLK_DISP1_SYS_PWR_REG when
enter low power mode(LPA and SLEEP). It changed source clock of the
ACLK_300_DISP1 from CPLL to XXTI for power saving. So,
the source clock of ACLK_300_DISP1 restore to CPLL by set as '1' to
CMU_SYSCLK_DISP1_SYS_PWR_REG when wakeup from low power mode.

Change-Id: Id3e55a9b9482b1990f22b6f8bc8e584d6330b587
Signed-off-by: Taekki Kim <taekki.kim@samsung.com>