devfreq: exynos5_bus_mif: set voltage to max during suspend
During resume, the bootloader may increase the frequency above
the maximum frequency the kernel uses.  If the kernel sets the
voltage to a value lower than the maximum voltage, the mif bus
may operate out of spec.

When entering suspend, set the voltage to the maximum possible
voltage, and the frequency to the maximum allowed frequency.

To minimize the time spent at maximum voltage, move the suspend
operation to a suspend handler instead of a pm notifier.

To minimize the time spent running at a frequency higher than
requested by the kernel, move the resume operation to a
resume_noirq instead of a pm notifier, and set the clock
frequency back to the maximum allowed frequency.

Also add a resume handler to put the frequency and voltage back
to what they were before suspend.

Change-Id: Ic0bcfc3cc56b7b020d20e12b6f91629d98dba1db
Signed-off-by: Colin Cross <ccross@android.com>