ARM: EXYNOS: pm: Change the location of exynos5_pmu_sysclk_save function.
If the first time try to suspend, suspend is aborted after syscore ops
but before s3c_pm_enter, the value never have been saved and
exynos_pm_resume restore it with 0. It's not symmetric with
exynos_pm_prepare. So, the location of exynos5_pmu_sysclk_save move to
exynos_pm_suspend.

Change-Id: Ia66b590d45d860dfd409056477e322a0f2d17f17
Signed-off-by: Taekki Kim <taekki.kim@samsung.com>