regulator: tps65090: Mark FET registers as volatile
The FET registers include both status and control bits.

Due to the regcache, we won't see changes in these registers since we
never read from the chip. Make the FET registers volatile to fix this.

One symptom of this problem is regultor error messages in the logs:

[    0.643077] s3c-i2c s3c2440-i2c.4: bus frequency set to 65 KHz
[    0.649217] tps65090 4-0048: No cache defaults, reading back from HW
[    0.771339] tps65090 4-0048: Disabling irq acks to chip.
[    0.796539] vcd_led: 12000 mV
[    0.811323] regulator regulator.20: reg 0x10 enable failed
[    0.816727] video_mid: failed to enable
[    0.826323] regulator regulator.21: reg 0x11 enable failed
[    0.831725] wwan_r: failed to enable
[    0.835442] sdcard: 3300 mV
[    0.846323] regulator regulator.23: reg 0x13 enable failed
[    0.851723] camout: failed to enable
[    0.855442] lcd_vdd: 3300 mV
[    0.858472] ts: 5000 mV
[    0.861067] vinchg:
[    0.863633] s3c-i2c s3c2440-i2c.4: i2c-4: S3C I2C adapter

I believe this is ultimately harmless, since we print a warning and continue.
Still, it is confusing.

BUG=chrome-os-partner:14576
TEST=manual
Boot snow kernel and see that regulator error messages no longer appear.

Change-Id: I181e8f7ed15c1df527fcfeb683cb43566002957f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34127
Reviewed-by: Doug Anderson <dianders@chromium.org>