OMAP4: PANDA: setup non configured GPIOs for LEDs and TFP410
In case TFP410 (Panda DPI to DVI) panel and leds_gpio are not enabled,
GPIOs have to be configured to a default setup to avoid random
peripheral behavior and useless power consumption.

Duplicate entries for Panda and PandaES for read-ability and easier
future expansion if required (the LED1 is different between these
boards).

This patch relies on the fact that the gpio_free will leave the
setup direction and value attributes of the GPIO.

Signed-off-by: Sebastien Jan <s-jan@ti.com>