gpu: t6xx: Per-context kernel-side memory free-list.
Zeroing of memory have proven to be slowing down some parts
of the GPU driver. To avoid zeroing of memory without
leaking information from either the kernel or from other
contexts, a per-context free-list is needed.

This patch removes a lot of code needed to support multiple
allocators, which were only needed if multiple source for
memory was availble. As this is not the case on any customer
board seen, it is best to remove this indirection.

Having removed the indirection a new memory allocator was
written which uses a free list to avoid zeroing memory.

Conflicts:
	drivers/gpu/arm/t6xx/kbase/mali_base_kernel.h

Bug: 7238983
Change-Id: I08d74eb0bc580d93469e424154840e255cf18056
Signed-off-by: Erik Gilling <konkers@android.com>