DSP: hwmod: don't use HW_AUTO to go to OFF mode
when increasing the DSP mmu latency constraint to be able to go to OFF,
the HW_AUTO configuration is crashing the DSP and generating L3 errors.
The current workaround is to force the hwmod framework to use SW_WKUP when
DSP powerdomain is ON and SW_SLEEP before moving the powerdomain to OFF.

To be able to go to suspend, we increase the mmu latency constraint to 1000.