gpu: t6xx: Correctly update core state when suspending
The coarse-mode power policy didn't update the availability
bitmap when it turned the GPU off. So once the GPU was
turned on the PM core would think the shader cores were
powered when in reality they were not.

This patch causes the PM core to update the bitmap when the
coarse-mode policy turnes off the GPU to indicate the true state of
the hardware.

Bug: 7259252
Change-Id: I477211b1770b2496579c742a552d4bd0a3c8efc9
Signed-off-by: John Einar Reitan <john.reitan@arm.com>