ARM: EXYNOS: pm_domain: Add save and restore for the pclk div consideration.
The pclk div of GSCL, MFC and DISP1 are placed on local power domain,
while its mirror registers are placed on top block. To update div
registers from mirror registers, mirror registers should be updated when
turn on the power domain. When turn on after turn off the power domain,
the pclks are bypass. It have possibility of wrong operation because
suppied the fast clock into the slow circuit.

Change-Id: Id843ce3444560bedbc9219f549cbc13849ae0a62
Signed-off-by: Taekki Kim <taekki.kim@samsung.com>