CHROMIUM: ALSA: ASoC: max98095 - Don't enable VDDPU by default.
The DSP clock selection was also setting bit four of the config register
which enables a pullup resistor between AVDD and JACKSNS.  The datasheet
specifies that this should not be set at the same time as MICBIAS2.
Setting both results in noise on MIC2, which is obvious on Daisy without
this fix.

BUG=chrome-os-partner:14903
TEST=arecord with headset listen for noise.  i2cdump shows register 45
doesn't have VDDPU set.

Change-Id: I5aa0e8be2652e550fef0942f16ea9214fe9571d5
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34679
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Tested-by: Evan Ragsdale <evan.ragsdale@maximintegrated.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>