mali: Per-context freelist max size
This patch adds support to set a soft max_size for the kbase
per-context freelist pools. The limit is soft in that the max might be
exceeded somewhat in some (safe) race conditions and might also limit
the size of the pool to be a bit lower than the max too in other
(safe) race conditions.

A hardcoded 64MB max size has been chosen for testing.

Bug: 7284501
Change-Id: Id75add9a8f0dccebdffdfc5bafce259083d1803d
Signed-off-by: John Einar Reitan <john.reitan@arm.com>