mali: vmalloc the kbase context structure
As the kbase context structure is large (~64kB) and kmalloc memory is
physically contiguous memory, creating a context depends on that a
large physically contiguous chunk exists. On a high-pressure/fragmented
system that might not be the case and the kbase context creation would
fail. vmalloced memory instead is built up of one or more physical
pages and better handles high-pressure/fragmented systems.

Bug: 7284501
Change-Id: I4d8d72ea7ccd61086b03ef7c78cb4d4a767bab1d
Signed-off-by: John Einar Reitan <john.reitan@arm.com>