gpu: t6xx: return correct value from pm_callback_power_on
After a power cycle of the GPU the chicken bits will have become unset.
The callback to power up the GPU needs to notify the driver that this
has happened so that the necessary chicken bits can be set again.
In case of GPU cycled, return 1 and If GPU is powered on, return 0 from
pm_callback_power_on.

bugs: 7243454, 7288104

Change-Id: I1b2ffa0f6f9983a9143e20771f32c93eaf610a6d
Signed-off-by: Seho Kim <sh1645.kim@samsung.com>