gpu: t6xx: Use mutex instead of a spinlock in freelist
As the time spent inside the free list allocation routine can take a
while when zeroing of pages is needed, holding a spinlock, which
blocks preemption, is unacceptable.

As the freelist code is never called from atomic context a mutex is a
better lock type.

This patch also defer the zeroing of pages to outside of the mutex
protected path.

Bug: 7315681
Change-Id: Ic72f256c04ab74dd41c3eb28668c9c4efe3362ea
Signed-off-by: John Einar Reitan <john.reitan@arm.com>