CHROMIUM: block: Remove deadlock in disk_clear_events
In disk_clear_events, do not put work on system_nrt_freezable_wq. Instead, put
it on system_nrt_wq.

There is a race between probing a usb and suspending the device. Since probing a
usb calls disk_clear_events, which puts work on a frozen workqueue, probing
cannot finish after the workqueue is frozen. However, suspending cannot finish
until the usb probe is finished, so we get a deadlock.

BUG=chrome-os-partner:14892, chromium-os:35228
TEST=suspend resume with usb device plugged in

Change-Id: Ife0103207c84eab162c93781d0d986eae50475d8
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/35324
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>