ARM: hw_breakpoint: v7.1 self-hosted debug powerdown support
This patch introduces os save and restore mechanism for v7.1 debug and
self-hosted debuggers.
It enables the os to save DBGDSCR before powerdown and restore it when power
is restored.

The clearing of the os lock in the restore function kick-starts the debug
logic again.

The os save and restore routines are hooked into the CPU PM event notifier
chain. CPU PM events are used to save and restore per-cpu context when a
single CPU is preparing to enter or has exited a low power state.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>