CHROMIUM: ASoC: samsung: fix error for hdmi-audio
The result of the device tree look-up for the hdmi-audio device
was checked against the wrong variable.
As a consequence, if the hdmi-audio device is disabled, the kernel will
crash at startup instead of skipping the plugin.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:14489
TEST=boot on Spring with hdmi-audio disabled in the device tree
and see we are longer crashing in the max98095 probing.

Change-Id: I613b53633eb7814e63a7459704f3780ded62a2a0
Reviewed-on: https://gerrit.chromium.org/gerrit/38261
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>