mmc: dw_mmc: fix to check data busy on request func by CMD13
CMD13 is used to check the card state whether card is busy or not
after block write. If busy time is long, CMD13 should be issued many times.
Instead of that case, old patch was added to reduce cpu load.
But it was not considered for some erase operations which takes long
times over a few seconds. If it was taking time over a few seconds,
though it returned error, it was not error.
This patch fixes CMD13 is able to issue every 1 seconds.

Bug: 7336393, 7118793

Change-Id: I82ae59e0ebb0a07b9660e509ed848dfe73333255
Signed-off-by: Hosung Kim <hosung0.kim@samsung.com>