mali: Missing cache flush in kbase free list mgmt
When a page on the free list is a low-mem page, the page itself is
used to store the metadata. High-mem uses a slab for this.
If we allocate a low-mem page from the free list we unlink it from the
free list. If the same page has been configured for dont-zero, then we
will have a dirty cache line due to the linked list unlink operation.
This must be flushed out before the page is handed out.

Bug: 7264664
Change-Id: I3b9a04747e67376dd16a193e2d7fb900d3f9b24c
Signed-off-by: John Einar Reitan <john.reitan@arm.com>