PM / Sleep: Add the concept of suspend_volatile
In order to catch suspend/resume problems we would like to run a CRC
over most of memory before and after sleep and compare the results.
To do this we need to know which memory regions the suspend/resume
code touches and avoid running the CRC over those regions.

We will consider those regions to be "suspend volatile".  The CRC code
will skip chunks of memory tagged as suspend volatile to avoid falsely
reporting CRC errors.

We add the ability to mark global variables as suspend volatile
(keeping them in a special section) and also the ability to register
dynamic chunks of memory as volatile.

BUG=chrome-os-partner:15914
TEST=suspend_stress_test

Change-Id: I7964f5886d6b62fbca8d82986db20eae83ad0918
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/37385
Reviewed-by: Michael Spang <spang@chromium.org>
Reviewed-by: Jon Kliegman <kliegs@chromium.org>