CHROMIUM: gobi: Move qcusbnet_put locking into free_dev
If the qcusbnet struct is freed through kobject_put instead of
qcusbnet_put, then we're not taking the lock that protects qcusbnet_list.
That's currently only possible if the struct is freed in cdev_put.

This bug was introduced in ae884f3 ("gobi: Keep struct qcusbnet alive
while cdev is alive").

BUG=chrome-os-partner:15849
TEST=suspend_stress_test

Change-Id: Ic34f7132e0e1a5724f02149b33f45ca71fa876f1
Signed-off-by: Michael Spang <spang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38078
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>