CHROMIUM: mfd/max77686: enable low-jitter mode
This change adds three valid device-tree configured settings
for the max77686 "low jitter" mode:
* max77686,enable-low-jitter
* max77686,disable-low-jitter
* (no entry)

With either entry, the third bit of the 32KHZ register
is updated.

Low jitter mode is desirable on some platforms with components that are
sensitive to the clock signal.  The clock sometimes appears to stutter when
being updated.  For most uses, this is inconsequential, but if the RTC is
supplying a signal to more sensitive hardware, it can cause undesirable
results.

BUG=chrome-os-partner:15785
TEST=It boots :) semenzato@ has confirmed that if the register is set, the errors do not reproduce with the scope.

To check the register state:
  $ i2cget -f -y 0 0x9 0x7f
  0x0f

Reported-by: Luigi Semenzato <semenzato@chromium.org>
Suggested-by: Luigi Semenzato <semenzato@chromium.org>
Signed-off-by: Will Drewry <wad@chromium.org>

Change-Id: Ib40d1d0728535b10c8438eaa6eacd74fb83cb697
Reviewed-on: https://gerrit.chromium.org/gerrit/38180
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Ready: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>