drm/exynos: Add drm notification of hotplug to dp
Instead of polling the connector for hotplug, use the hotplug workqueue
to fire a drm hotplug notification when we detect a cable change.

This will also update userspace when the state has changed, and notify
fbdev.

At this point in time, the patch should be a no-op, but it's important
if we want to use EDID to mode set.

BUG=chrome-os-partner:11158
TEST=Tested on snow, no regressions detected

Change-Id: Ib6b0c789f6abc45a118a04dcb8a73e8e35fd4a85
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38408
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>