drm/exynos: Clean up fimd mode handling
FIMD has been using the panel timing data straight from the driver's
platform data instead of the data that is provided via the drm layer.
This makes things difficult if we want to use EDID data from the dp
driver, since it would be ignored. This patch changes FIMD to just
use the platform panel data as a seed to get_panel. From then on, we use
the overlay data that is set via modeset.

This patch also removes the panel array, since we only want one
resolution from platform data.

BUG=chrome-os-partner:11158
TEST=Tested on snow, no regressions noted

Change-Id: I345c3663509e6820a745b85c4b9a37f77f42c0d2
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38409
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>