CHROMIUM: rts_pstor: prevent deadlock and detection issues
Patch from Realtek that fixes a deadlock issue when combined with removing the
sdhci_pci module from the kernel. This disables the card interrupt over the bus
and something else that's undocumented.

The change also prevents issues with the card reader either not detecting an SD
card insertion or removal.

TEST=suspend, insert sd card, resume
BUG=chrome-os-partner:14934, chrome-os-partner:16053

Change-Id: Ibe81ab819662b8d6721dcfe5357d12eaf2a2335d
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38713
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>